Initial version of event client
